slack/pyproject.toml
[project]
name = "weave-bot-slack"
version = "0.1.0"
description = "Slack bot for weave-bot-orb event discovery"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
